+++ mqttwarn/services/prowl.py
@@ -2,10 +2,10 @@
 # -*- coding: utf-8 -*-
 
 __author__    = 'Jan-Piet Mens <jpmens()gmail.com>'
-__copyright__ = 'Copyright 2014 Jan-Piet Mens'
+__copyright__ = 'Copyright 2014-2021 Jan-Piet Mens'
 __license__   = 'Eclipse Public License - v 1.0 (http://www.eclipse.org/legal/epl-v10.html)'
 
-import prowlpy
+import pyprowl
 
 
 def plugin(srv, item):
@@ -17,18 +17,23 @@ def plugin(srv, item):
 
     title = item.get('title', srv.SCRIPTNAME)
     text = item.message
-    priority = item.get('priority', 0)
+    priority = int(item.get('priority', 0))
 
     try:
-        p = prowlpy.Prowl(apikey)
-        p.post(application=title,
-            event=application,
-            description=text,
-            priority=priority,
-            providerkey=None,
-            url=None)
+        p = pyprowl.Prowl(apikey)
+        p.verify_key()
+        srv.logging.info("Prowl API key successfully verified")
     except Exception as e:
-        srv.logging.warning("Cannot prowl: %s" % e)
+        srv.logging.error("Error verifying Prowl API key: {}".format(e))
+        return False
+
+    try:
+        p.notify(event=title, description=text,
+                 priority=priority, url=None,
+                 appName=application)
+        srv.logging.debug("Sending notification to Prowl succeeded")
+    except Exception as e:
+        srv.logging.warning("Sending notification to Prowl failed: %s" % e)
         return False
 
     return True